Fixed bug causing wrong order of cartridges on arrow down. Refactored BrickGame component, logic extracted to a hook
